parsing norminette finished
