+++ include/minishell.h
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   minishell.h                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kglebows <kglebows@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/01/19 10:28:06 by kglebows          #+#    #+#             */
+/*   Updated: 2024/01/19 13:12:20 by kglebows         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef MINISHELL_H
 # define MINISHELL_H
 
@@ -14,27 +26,17 @@
 # include <termios.h>
 # include <readline/readline.h>
 # include <readline/history.h>
-
-// FOR LINUX VALGRIND
-// # include <wait.h>
-//
-
 # include "../lib/libft/libft.h"
-
+# include "struct.h"
 # include "parse.h"
 # include "builtin.h"
 # include "executor.h"
 # include "utils.h"
+# include "exit.h"
 
-
-// void	ft_executor(t_dt *dt);
-t_return	ft_parse(t_dt *dt);
-t_env	*create_env_var_node(char *str);
 void	setup_signals(void);
 void	setup_child_signals(void);
-
-//signal_handler
-// void	init_signal_handler(void);
+void	block_signal(void);
 
 # define FALSE 0
 # define TRUE 1
