+++ src/executor/exe.c
@@ -48,6 +48,7 @@ void	execute(t_cmdtable *table, t_dt *minishell, bool last_cmd)
 
 	if (pipe(fd) == -1)
 		ft_putstr_fd("Pipe Error\n", 2);
+	// printf("1: %d 2: %d\n", fd[0], fd[1]);
 	if (table->cmd[0])
 		exe_parent_builtin_cmds(table, minishell);
 	block_signal();
@@ -67,9 +68,16 @@ void	execute(t_cmdtable *table, t_dt *minishell, bool last_cmd)
 	{
 		table->fd_rdr_out = 0;
 		close(fd[1]);
-		dup2(fd[0], STDIN_FILENO);
 		if (last_cmd)
+		{
 			dup2(table->fd_in, STDIN_FILENO);
+			close(fd[0]);
+		}
+		else
+		{
+			dup2(fd[0], STDIN_FILENO);
+			close(fd[0]);
+		}
 	}
 }
 
@@ -79,19 +87,33 @@ void	child(t_cmdtable *table, bool last_cmd, t_env *envp_lst, int *fd)
 
 	close(fd[0]);
 	if (check_redirections(table, fd) && table->cmd)
+	{
+		// TODO: here
 		exit(1);
+	}
 	else
 	{
 		dup2(fd[1], STDOUT_FILENO);
+		close(fd[1]);
 		if (table->fd_rdr_out)
+		{
 			dup2(table->fd_rdr_out, STDOUT_FILENO);
-		if (last_cmd && !table->fd_rdr_out)
+			close(table->fd_rdr_out);
+		}
+		else if (last_cmd && !table->fd_rdr_out)
+		{
 			dup2(table->fd_out, STDOUT_FILENO);
+			close(table->fd_out);
+		}
 		if (!exe_built_in_cmds(table->cmd, envp_lst))
+		{
+			// TODO: here
 			exit(EXIT_SUCCESS);
+		}
 		else
 		{
 			path = find_path(table, env_to_char_array(envp_lst), last_cmd);
+			// TODO: here
 			execve(path, table->cmd, env_to_char_array(envp_lst));
 		}
 	}
