+++ src/exit/error.c
@@ -1,55 +1,66 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   error.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kglebows <kglebows@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/01/19 17:07:34 by kglebows          #+#    #+#             */
+/*   Updated: 2024/01/19 17:07:37 by kglebows         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "minishell.h"
 
-void	error_(int code)
+void	error_(int code, t_dt *dt)
 {
+	(void)dt;
 	if (code == -10)
-		printf("minishell: Memory allocation failed!\n");
+		ft_printf("minishell: Memory allocation failed!\n");
 	else if (code == -11)
-		printf("minishell: Quotes not closed!\n");
+		ft_printf("minishell: Quotes not closed!\n");
 	else if (code == -12)
-		printf("minishell: syntax error near unexpected token `|'\n");
+		ft_printf("minishell: syntax error near unexpected token `|'\n");
 	else if (code == -13)
-		printf("minishell: syntax error near unexpected token\n");
+		ft_printf("minishell: syntax error near unexpected token\n");
 	else if (code == -14)
-		printf("minishell: syntax error near unexpected token\n");
+		ft_printf("minishell: syntax error near unexpected token\n");
 	else if (code == -15)
-		printf("Error! \n");
+		ft_printf("Error! \n");
 	else if (code == -16)
-		printf("Error! \n");
+		ft_printf("Error! \n");
 	else if (code == -17)
-		printf("Error! \n");
+		ft_printf("Error! \n");
 	else if (code == -18)
-		printf("Error! \n");
+		ft_printf("Error! \n");
 	else if (code == -19)
-		printf("Error! \n");
+		ft_printf("Error! \n");
 	else
-		printf("Error! Unknown error code %d!\n", code);
+		ft_printf("Error! Unknown error code %d!\n", code);
 }
 
-t_return	ft_error(int code, t_dt *dt)
+t_ok	ft_error(int code, t_dt *dt)
 {
 	if (code == -1)
-		printf("minishell: Minishell takes no arguments!\n");
+		ft_printf("minishell: Minishell takes no arguments!\n");
 	else if (code == -2)
-		printf("Error! \n");
+		ft_printf("Error! \n");
 	else if (code == -3)
-		printf("Error! \n");
+		ft_printf("Error! \n");
 	else if (code == -4)
-		printf("Error! \n");
+		ft_printf("Error! \n");
 	else if (code == -5)
-		printf("Error! \n");
+		ft_printf("Error! \n");
 	else if (code == -6)
-		printf("Error! \n");
+		ft_printf("Error! \n");
 	else if (code == -7)
-		printf("Error! \n");
+		ft_printf("Error! \n");
 	else if (code == -8)
-		printf("Error! \n");
+		ft_printf("Error! \n");
 	else if (code == -9)
-		printf("Error! \n");
+		ft_printf("Error! \n");
 	else
-		error_(code);
-	(void)dt;
+		error_(code, dt);
 	if (code > -10)
 		exit_code(2);
 	else
