+++ src/parse/env.c
@@ -1,133 +1,116 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   env.c                                              :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kglebows <kglebows@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/01/19 10:12:16 by kglebows          #+#    #+#             */
+/*   Updated: 2024/01/19 10:32:30 by kglebows         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "minishell.h"
 
-/**
- * @brief Adds or Changes a value in Envp list. if key exists, frees memory of key parameter.
- * @param key pointer to allocated string with key
- * @param value pointer to allocated string with value
- * @param dt main data structure
- * @return KO if creation of new node failed or key is NULL
-*/
-t_return	envp_add_or_change(char *key, char *value ,t_dt *dt)
+char	*ft_strgetbetween(const char *start, const char *end)
 {
-	t_env		*temp;
-	t_env		*last;
+	char	*word;
+	size_t	i;
 
-	if (!key)
-		return (KO);
-	temp = dt->envp_lst;
-	while (temp)
+	if ((*start == '\'' && *end == '\'') || (*start == '"' && *end == '"'))
 	{
-		if (ft_strncmp(key, temp->key, ft_strlen(key)) == 0
-			&& ft_strlen(key) == ft_strlen(temp->key))
-			break ;
-		last = temp;
-		temp = temp->next;
+		start++;
+		end--;
 	}
-	if (temp)
+	word = (char *)malloc((end - start + 2) * sizeof(char));
+	if (!word)
+		return (NULL);
+	i = 0;
+	while (*start && start <= end)
 	{
-		free(key);
-		free(temp->value);
-		temp->value = value;
-		return (OK);
+		word[i] = *start;
+		start++;
+		i++;
 	}
-	last->next = ft_calloc(1, sizeof(t_env));
-	if (!last->next)
-		return(ft_error(-10, dt));
-	last->next->key = key;
-	last->next->value = value;
-	last->next->next = NULL;
-	return (OK);
+	word[i] = '\0';
+	return (word);
 }
 
-/**
- * @brief Removes an entry of key, if no key found - does nothing.Frees input key.
- * @param key pointer to allocated string with key
- * @param dt main data structure
- * @return KO if key NULL or missing, OK if entry deleted.
-*/
-t_return	envp_delete(char *key, t_dt *dt)
+t_env	*create_env_var_node_(char *str, t_env	*node, char	*eq_sign_pos)
 {
-	t_env		*temp;
-	t_env		*last;
+	node->next = NULL;
+	node->key = ft_strgetbetween(str, eq_sign_pos - 1);
+	if (!node->key)
+		return (free(node), NULL);
+	node->value = ft_strdup(eq_sign_pos + 1);
+	if (!node->value)
+		return (free(node->key), free(node), NULL);
+	return (node);
+}
 
-	if (!key)
-		return (KO);
-	temp = dt->envp_lst;
-	while (temp)
+t_env	*create_env_var_node(char *str)
+{
+	char	*eq_sign_pos;
+	t_env	*node;
+
+	eq_sign_pos = ft_strchr(str, '=');
+	if (!eq_sign_pos)
 	{
-		if (ft_strncmp(key, temp->key, ft_strlen(key)) == 0
-			&& ft_strlen(key) == ft_strlen(temp->key))
-			break ;
-		last = temp;
-		temp = temp->next;
+		node = (t_env *)ft_calloc(1, sizeof(t_env));
+		if (!node)
+			return (NULL);
+		node->next = NULL;
+		node->key = ft_strdup(str);
+		if (!node->key)
+			return (free(node), NULL);
+		node->value = NULL;
+		return (node);
 	}
-	if (temp)
+	node = (t_env *)ft_calloc(1, sizeof(t_env));
+	if (!node)
+		return (NULL);
+	return (create_env_var_node_(str, node, eq_sign_pos));
+}
+
+void	free_env_var_list(t_env *head)
+{
+	t_env	*temp;
+
+	while (head != NULL)
 	{
-		last->next = temp->next;
+		temp = head;
+		head = head->next;
 		free(temp->key);
 		free(temp->value);
 		free(temp);
-		free(key);
-		return (OK);
 	}
-	free(key);
-	return (KO);
 }
 
-/**
- * @brief Prints envp list for export command. it shows all declared variables, even those without value
- * @param dt main data structure
-*/
-void	envp_print_export(t_dt *dt)
+t_env	*init_env_var(char *envp[])
 {
-	t_env			*temp;
+	t_env	*head;
+	t_env	*tail;
+	t_env	*new_node;
+	int		i;
 
-	temp = dt->envp_lst;
-	while (temp)
+	head = NULL;
+	tail = NULL;
+	i = 0;
+	while (envp[i++] != NULL)
 	{
-		ft_printf("declare -x %s", temp->key);
-		if (temp->value)
-			ft_printf("=\"%s\"\n", temp->value);
+		new_node = create_env_var_node(envp[i - 1]);
+		if (new_node == NULL)
+			return (free_env_var_list(head), NULL);
+		if (tail == NULL)
+		{
+			head = new_node;
+			tail = new_node;
+		}
 		else
-			ft_printf("\n");
-		temp = temp->next;
+		{
+			tail->next = new_node;
+			tail = new_node;
+		}
 	}
+	return (head);
 }
-/**
- * @brief Prints envp list for env command. it shows only entries that have value
- * @param dt main data structure
-*/
-void	envp_print_env(t_dt *dt)
-{
-	t_env			*temp;
-
-	temp = dt->envp_lst;
-	while (temp)
-	{
-		if (temp->value)
-			ft_printf("%s=%s\n", temp->key, temp->value);
-		temp = temp->next;
-	}
-}
-
-/*
-	env, unset and export implementation :
-		env
-			run	env_print_env(dt);
-		unset
-			run	envp_delete(NULL, dt);
-		unset key
-			run	envp_delete(key, dt);
-		unset key1 key2 key3 ... 
-			run	envp_delete(key1, dt);
-			run	envp_delete(key2, dt);
-			run	envp_delete(key3, dt);
-		export
-			run	env_print_export(dt);
-		export key || export key=value
-			allocate memory for key
-			allocate memory for value (if value != NULL)
-			run	envp_add_or_change(key, value , dt);
-		export key1 key2=value2 key3 ...
-			repeat {export key || export key=value} for every argument
-*/
