+++ src/parse/expander.c
@@ -1,3 +1,14 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   expander.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kglebows <kglebows@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/01/19 10:47:35 by kglebows          #+#    #+#             */
+/*   Updated: 2024/01/19 17:11:50 by kglebows         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "minishell.h"
 
@@ -22,38 +33,6 @@ char	*check_env(char *check, int i, t_dt *dt)
 	}
 	return (ret);
 }
-char	*ft_strndup(const char *s, size_t n)
-{
-	char *result = malloc(n + 1);
-	if (result) {
-		strncpy(result, s, n);
-		result[n] = '\0';
-	}
-	return result;
-}
-// char	*expand(char q, char *start, t_dt *dt)
-// {
-// 	int		i;
-// 	char	*temp;
-// 	char	*exp;
-
-// 	i = 0;
-// 	if (start[i] == '?')
-// 		return (ft_itoa(*ini_exit()));
-// 	while (start[i] == '_' || ft_isalnum(start[i]) == 1)
-// 		i++;
-// 	if (i == 0)
-// 		return (ft_calloc(1, sizeof(char)));
-// 		// return (ft_strdup("$")); // Return "$" when there is no alphanumeric character following the '$'
-// 	temp = ft_strndup(start, i);
-// 	if (!temp)
-// 		ft_error(-10, dt);
-// 	if (q == '\'')
-// 		return (temp);
-// 	exp = check_env(temp, i, dt);
-// 	free(temp);
-// 	return (exp ? exp : ft_strdup(""));
-// }
 
 char	*expand(char q, char *start, t_dt *dt)
 {
@@ -73,8 +52,6 @@ char	*expand(char q, char *start, t_dt *dt)
 	if (i == 0)
 		return (ft_calloc(1, sizeof(char)));
 	temp = calloc(i + 1, sizeof(char));
-	if (!temp)
-		ft_error(-10, dt);
 	if (q == '\'')
 	{
 		ft_strlcpy(temp, &start[-1], i + 2);
@@ -101,40 +78,40 @@ char	*update_string(char *str, char *update)
 	return (temp);
 }
 
-char	*ft_expander(char *str, int size, t_dt *dt)
+int	ft_expander_(char *str, int i)
+{
+	if (str[i + 1] == '?')
+		i++;
+	else
+	{
+		while (str[i + 1] == '_' || ft_isalnum(str[i + 1]) == 1)
+			i++;
+	}
+	return (i);
+}
+
+char	*ft_expander(char *s, int size, t_dt *dt)
 {
 	char	q;
 	char	*temp;
 	int		i;
-	char	*c;
 
 	q = ' ';
 	i = 0;
 	temp = ft_calloc(1, sizeof(char));
 	while (i < size)
 	{
-		if (q == ' ' && (str[i] == '\'' || str[i] == '\"'))
-			q = str[i];
-		else if ((q == '\'' && str[i] == '\'') || (q == '\"' && str[i] == '\"'))
+		if (q == ' ' && (s[i] == '\'' || s[i] == '\"'))
+			q = s[i];
+		else if ((q == '\'' && s[i] == '\'') || (q == '\"' && s[i] == '\"'))
 			q = ' ';
-		else if (str[i] == '$')
+		else if (s[i] == '$')
 		{
-			temp = update_string(temp, expand(q, &str[i + 1], dt));
-			if (str[i + 1] == '?')
-				i++;
-			else
-			{
-				while (str[i + 1] == '_' || ft_isalnum(str[i + 1]) == 1)
-					i++;
-			}
+			temp = update_string(temp, expand(q, &s[i + 1], dt));
+			i = ft_expander_(s, i);
 		}
 		else
-		{
-			c = ft_calloc(2, sizeof(char));
-			c[0] = str[i];
-			c[1] = '\0';
-			temp = update_string(temp, c);
-		}
+			temp = update_string(temp, ft_onecharstring(s[i]));
 		i++;
 	}
 	if (q != ' ')
